# docs/api/openapi.yml
openapi: 3.1.0
info:
  title: 数字小窝 API
  version: 0.1.0
  description: 宠物数字纪念网站当前可用接口。所有写操作需要同域 XSRF Cookie 与 X-XSRF-TOKEN 请求头；公开页的口令解锁也适用。
servers:
  - url: /api/v1
paths:
...
  /custom-service-requests:
    get:
      summary: 获取当前用户的人工定制申请
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: 当前用户的服务申请列表 }
    post:
      summary: 提交人工定制申请
      description: 仅登记需求与联系方式，不创建订单、不扣款；实际收费须在正式商户支付流程开放后另行确认。
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCustomServiceRequest" }
      responses:
        "201": { description: 申请已提交 }
...
  /admin/overview:
    get:
      summary: 获取运营管理概览
      description: 返回审核、订单、匿名访问与受控媒体字节用量汇总。访问按纪念页、匿名访客指纹和日期去重；服务端不保存原始 IP 地址，也不在概览中返回媒体内容。
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: 管理概览，仅 ADMIN 或 MODERATOR 可访问 }
        "403": { description: 缺少运营角色 }
  /admin/tributes:
    get:
      summary: 获取留言审核队列
      security: [{ sessionCookie: [] }]
      parameters:
        - name: status
          in: query
          schema:
            {
              type: string,
              enum: [PENDING, REPORTED, APPROVED, HIDDEN, ALL],
              default: PENDING,
            }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200": { description: 留言审核记录 }
  /admin/tributes/{tributeId}:
    patch:
      summary: 审核或隐藏一条留言
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/TributeId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ModerateTributeRequest" }
      responses:
        "200": { description: 留言审核完成并写入审计日志 }
        "403": { description: 缺少运营角色 }
  /admin/orders:
    get:
      summary: 获取订单概览
      security: [{ sessionCookie: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200": { description: 订单概览，仅 ADMIN 或 MODERATOR 可访问 }
  /admin/orders/{orderId}/mock-refund:
    post:
      summary: 撤销一笔本地模拟支付
      description: 仅在服务端显式开启本地模拟支付时可用。运营角色必须填写原因；订单和对应权益在同一事务中更新，重复请求不会重复撤销或新增审计记录。正式支付退款必须通过支付渠道接口处理。
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/AdminOrderId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MockRefundRequest" }
      responses:
        "200": { description: 本地模拟退款结果 }
        "403": { description: 缺少运营角色 }
        "404": { description: 订单不存在 }
        "409": { description: 模拟支付未启用或订单当前不可退款 }
  /admin/memorials:
    get:
      summary: 获取公开纪念页巡查队列
      security: [{ sessionCookie: [] }]
      parameters:
        - name: status
          in: query
          schema:
            {
              type: string,
              enum: [PUBLISHED, ARCHIVED, ALL],
              default: PUBLISHED,
            }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200": { description: 纪念页巡查记录，仅 ADMIN 或 MODERATOR 可访问 }
  /admin/memorials/{memorialId}:
    patch:
      summary: 因内容审核下线一间公开纪念页
      description: 仅能将已发布页面下线，必须填写原因并写入审计日志；恢复分享仍由页面所有者完成。
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/AdminMemorialId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ArchiveMemorialRequest" }
      responses:
        "200": { description: 页面已下线并写入审计日志 }
        "403": { description: 缺少运营角色 }
        "404": { description: 纪念页不存在 }
        "409": { description: 页面当前未发布，无法下线 }
  /admin/custom-service-requests:
    get:
      summary: 获取人工定制运营队列
      security: [{ sessionCookie: [] }]
      parameters:
        - name: status
          in: query
          schema:
            {
              type: string,
              enum:
                [
                  OPEN,
                  ALL,
                  SUBMITTED,
                  MATERIALS_PENDING,
                  IN_PROGRESS,
                  REVISION,
                  OUT_OF_SCOPE,
                  DELIVERED,
                  COMPLETED,
                  CANCELED,
                ],
              default: OPEN,
            }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200": { description: 人工定制工单列表 }
  /admin/custom-service-requests/{requestId}:
    patch:
      summary: 更新人工定制运营工单
      description: 每次更新都要求填写原因，并写入审计日志；工作人员不能替用户确认交付完成。
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: "#/components/parameters/CustomServiceRequestId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCustomServiceRequest" }
      responses:
        "200": { description: 工单已更新 }
        "409": { description: 非法状态流转或需要用户确认交付 }
  /admin/audit-logs:
    get:
      summary: 获取运营操作审计日志
      security: [{ sessionCookie: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200": { description: 审计日志列表 }
  /media/images:
    post:
      summary: 上传纪念页图片
      description: 仅接受不超过 5MB 的 JPG、PNG 或 WebP。服务端同时核验声明的 MIME 类型与文件签名，不能仅靠改文件扩展名或请求头绕过格式限制。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201": { description: 图片已保存 }
        "413": { description: 图片超过 5MB }
        "415": { description: 不支持的文件格式 }
  /media/videos:
    post:
      summary: 上传纪念页短视频
      description: 仅接受不超过 30MB 的 MP4 或 WebM。服务端同时核验声明的 MIME 类型与容器签名；视频只作为相册媒体展示，不自动播放，且由套餐短视频数量限制控制。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201": { description: 短视频已保存 }
        "413": { description: 短视频超过 30MB }
        "415": { description: 不支持的文件格式或文件签名 }
  /media/{id}/content:
    get:
      summary: 读取受控媒体二进制内容
      description: 草稿媒体仅限所有者读取；公开或持链接页面引用的媒体可访问；口令页面需要先在同一浏览器会话解锁。短视频支持 HTTP byte range 读取以便拖动播放进度。
      parameters: [{ $ref: "#/components/parameters/MediaId" }]
      responses:
        "200": { description: 图片内容 }
        "404": { description: 图片不存在或无查看权限 }
  /memorials:
    post:
      summary: 创建私密纪念页草稿
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemorialRequest" }
      responses:
        "201": { description: 草稿已创建 }
...
  /memorials/{id}:
    get:
      summary: 获取自己的纪念页详情
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/MemorialId" }]
      responses:
        "200": { description: 纪念页详情 }
        "404": { description: 不存在或无权访问 }
    put:
      summary: 更新自己的纪念页
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/MemorialId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MemorialRequest" }
      responses:
        "200":
          {
            description: 已更新；已发布页面切换为仅自己可见时会自动下线，原分享链接立即不可访问。,
          }
...
  /memorials/{id}/tributes:
    get:
      summary: 获取自己纪念页的全部留言（含隐藏和待处理）
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/MemorialId" }]
      responses:
        "200": { description: 留言列表 }
  /memorials/{id}/tributes/{tributeId}:
    patch:
      summary: 显示或隐藏一条亲友留言
      security: [{ sessionCookie: [] }]
      parameters:
        - { $ref: "#/components/parameters/MemorialId" }
        - { $ref: "#/components/parameters/TributeId" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TributeStatusRequest" }
      responses:
        "200": { description: 留言状态已更新 }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
  parameters:
    MemorialId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    AdminMemorialId:
      name: memorialId
      in: path
      required: true
      schema: { type: string, format: uuid }
    AdminOrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
    MediaId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    TributeId:
      name: tributeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CustomServiceRequestId:
      name: requestId
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    CreateCustomServiceRequest:
      type: object
      required: [contactDetails, requestDetails]
      properties:
        contactDetails: { type: string, minLength: 3, maxLength: 280 }
        requestDetails: { type: string, minLength: 10, maxLength: 3000 }
    ModerateTributeRequest:
      type: object
      required: [status, reason]
      properties:
        status: { type: string, enum: [APPROVED, HIDDEN] }
        reason: { type: string, minLength: 2, maxLength: 280 }
    ArchiveMemorialRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 2, maxLength: 280 }
    MockRefundRequest:
      type: object
      required: [reason]
      properties:
        reason: { type: string, minLength: 2, maxLength: 280 }
    UpdateCustomServiceRequest:
      type: object
      required: [status, materialsReady, revisionCount, reason]
      properties:
        status:
          {
            type: string,
            enum:
              [
                SUBMITTED,
                MATERIALS_PENDING,
                IN_PROGRESS,
                REVISION,
                OUT_OF_SCOPE,
                DELIVERED,
                COMPLETED,
                CANCELED,
              ],
          }
        materialsReady: { type: boolean }
        assignee: { type: [string, "null"], maxLength: 80 }
        dueDate: { type: [string, "null"], format: date }
        revisionCount: { type: integer, minimum: 0, maximum: 50 }
        customerMessage: { type: [string, "null"], maxLength: 800 }
        reason: { type: string, minLength: 2, maxLength: 280 }
    MemorialRequest:
      type: object
      required: [petName, species, visibility, version]
      properties:
        petName: { type: string, maxLength: 32 }
        species: { type: string, maxLength: 32 }
        coverMediaId: { type: [string, "null"], format: uuid }
        farewellMessage: { type: [string, "null"], maxLength: 280 }
        aboutTa:
          {
            type: [string, "null"],
            maxLength: 1500,
            description: 宠物的性格、习惯、喜欢的事与昵称由来,
          }
        companionStartedOn:
          {
            type: [string, "null"],
            format: date,
            description: 可选的相遇、领养或出生日期,
          }
        companionEndedOn:
          {
            type: [string, "null"],
            format: date,
            description: 可选的想念开始或陪伴结束日期,
          }
        visibility: { type: string, enum: [PUBLIC, LINK, PASSWORD, PRIVATE] }
        theme: { type: string, enum: [SUNNY, NIGHT, GARDEN], default: SUNNY }
        accessCode: { type: [string, "null"], minLength: 6, maxLength: 64 }
        version: { type: integer, minimum: 0 }
...
    TributeStatusRequest:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [APPROVED, HIDDEN] }
...